+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for HmIP-MP3P (Combination Signalling Device)
 - Add extended smoke detector properties for HmIP-SWSD-2
 - Add support for HmIP-LSC (Light Strip Controller)
+- Add support for HmIP-USBSM (USB Switch Measuring)
+- Add support for ELV-SH-SB8 Status Board
 - Add support for ELV-SH-PSMCI (Switch Measuring Cable Indoor)
 - Add test coverage for HmIPW-DRI16 (Wired Input 16)